fix: UTC_TIME for RS128 is in "s.ns" format.
This commit fixs the error in decoder_RS128.hpp. For RS128, the timestamp data in the MSOP header is in "s.ns" format, instead of "s.us".
However, the old code treats the timestamp as in "s.us" format, which cause significantly error when parsing the UTC timestamp from the LiDAR.